Change: restructured landingpage to be loginpage.
