--- ManiFul/src/screens/LandingPage/index.ts
@@ -1,2 +1,0 @@
-//LandingPage index.ts
-export { default } from './LandingPage';
