+++ ManiFul/src/screens/LoginPage/index.ts
@@ -1,2 +1,2 @@
-//LoginPage index.ts
+//LandingPage index.ts
 export { default } from './LoginPage';
