+++ ManiFul/src/screens/LoginPage/styles.ts
@@ -1,9 +1,28 @@
-//LoginPage styles.ts
+//LandingPage styles.ts
 import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
   container: {
     flex: 1,
-    padding: 16,
+    justifyContent: 'flex-start',
+    alignItems: 'center',
+  },
+  box: {
+    //backgroundColor: 'white',
+    aspectRatio: 1,
+    marginTop: '10%',
+    width: '80%',
+    display: 'flex',
+    justifyContent: 'flex-start',
+    alignItems: 'center',
+    position: 'relative',
+  },
+  loginField: {
+    position: 'relative',
+    marginTop: 40,
+    display: 'flex',
+    width: '100%',
+    justifyContent: 'center',
+    alignItems: 'center',
   },
 });
