+++ ManiFul/src/types/navigation.ts
@@ -2,17 +2,10 @@
 import { NativeStackNavigationProp } from '@react-navigation/native-stack';
 
 export type RootStackParamList = {
-  landingPage: undefined;
   login: undefined;
   home: undefined;
   // Add other screens here
 };
-
-// Navigation prop type for LandingPage
-export type LandingPageNavigationProp = NativeStackNavigationProp<
-  RootStackParamList,
-  'landingPage'
->;
 
 export type LoginPageNavigationProp = NativeStackNavigationProp<
   RootStackParamList,
